[WFCORE-2657] - make mechanism-name verification work, add single class to validate string allowed values
