+++ elytron/src/main/java/org/wildfly/extension/elytron/JdbcRealmDefinition.java
@@ -51,10 +51,8 @@ import org.jboss.as.controller.SimpleAttributeDefinition;
 import org.jboss.as.controller.SimpleAttributeDefinitionBuilder;
 import org.jboss.as.controller.SimpleResourceDefinition;
 import org.jboss.as.controller.capability.RuntimeCapability;
-import org.jboss.as.controller.logging.ControllerLogger;
-import org.jboss.as.controller.operations.validation.AllowedValuesValidator;
+import org.jboss.as.controller.operations.validation.StringAllowedValuesValidator;
 import org.jboss.as.controller.operations.validation.IntRangeValidator;
-import org.jboss.as.controller.operations.validation.ModelTypeValidator;
 import org.jboss.as.controller.registry.AttributeAccess;
 import org.jboss.as.controller.registry.ManagementResourceRegistration;
 import org.jboss.as.controller.registry.OperationEntry;
@@ -98,7 +96,7 @@ class JdbcRealmDefinition extends SimpleResourceDefinition {
         static final SimpleAttributeDefinition ALGORITHM = new SimpleAttributeDefinitionBuilder(ElytronDescriptionConstants.ALGORITHM, ModelType.STRING)
                 .setRequired(false)
                 .setDefaultValue(new ModelNode(ClearPassword.ALGORITHM_CLEAR))
-                .setValidator(new StringValuesValidator(ClearPassword.ALGORITHM_CLEAR))
+                .setValidator(new StringAllowedValuesValidator(ClearPassword.ALGORITHM_CLEAR))
                 .setAllowExpression(false)
                 .setFlags(AttributeAccess.Flag.RESTART_RESOURCE_SERVICES)
                 .build();
@@ -147,7 +145,7 @@ class JdbcRealmDefinition extends SimpleResourceDefinition {
         static final SimpleAttributeDefinition ALGORITHM = new SimpleAttributeDefinitionBuilder(ElytronDescriptionConstants.ALGORITHM, ModelType.STRING)
                 .setRequired(false)
                 .setDefaultValue(new ModelNode(BCryptPassword.ALGORITHM_BCRYPT))
-                .setValidator(new StringValuesValidator(BCryptPassword.ALGORITHM_BCRYPT))
+                .setValidator(new StringAllowedValuesValidator(BCryptPassword.ALGORITHM_BCRYPT))
                 .setAllowExpression(false)
                 .setFlags(AttributeAccess.Flag.RESTART_RESOURCE_SERVICES)
                 .build();
@@ -211,7 +209,7 @@ class JdbcRealmDefinition extends SimpleResourceDefinition {
         static final SimpleAttributeDefinition ALGORITHM = new SimpleAttributeDefinitionBuilder(ElytronDescriptionConstants.ALGORITHM, ModelType.STRING)
                 .setRequired(false)
                 .setDefaultValue(new ModelNode(SaltedSimpleDigestPassword.ALGORITHM_PASSWORD_SALT_DIGEST_MD5))
-                .setValidator(new StringValuesValidator(
+                .setValidator(new StringAllowedValuesValidator(
                         SaltedSimpleDigestPassword.ALGORITHM_PASSWORD_SALT_DIGEST_MD5,
                         SaltedSimpleDigestPassword.ALGORITHM_PASSWORD_SALT_DIGEST_SHA_1,
                         SaltedSimpleDigestPassword.ALGORITHM_PASSWORD_SALT_DIGEST_SHA_256,
@@ -278,7 +276,7 @@ class JdbcRealmDefinition extends SimpleResourceDefinition {
         static final SimpleAttributeDefinition ALGORITHM = new SimpleAttributeDefinitionBuilder(ElytronDescriptionConstants.ALGORITHM, ModelType.STRING)
                 .setRequired(false)
                 .setDefaultValue(new ModelNode(SimpleDigestPassword.ALGORITHM_SIMPLE_DIGEST_MD5))
-                .setValidator(new StringValuesValidator(
+                .setValidator(new StringAllowedValuesValidator(
                         SimpleDigestPassword.ALGORITHM_SIMPLE_DIGEST_MD2,
                         SimpleDigestPassword.ALGORITHM_SIMPLE_DIGEST_MD5,
                         SimpleDigestPassword.ALGORITHM_SIMPLE_DIGEST_SHA_1,
@@ -333,7 +331,7 @@ class JdbcRealmDefinition extends SimpleResourceDefinition {
         static final SimpleAttributeDefinition ALGORITHM = new SimpleAttributeDefinitionBuilder(ElytronDescriptionConstants.ALGORITHM, ModelType.STRING)
                 .setRequired(false)
                 .setDefaultValue(new ModelNode(ScramDigestPassword.ALGORITHM_SCRAM_SHA_256))
-                .setValidator(new StringValuesValidator(
+                .setValidator(new StringAllowedValuesValidator(
                         ScramDigestPassword.ALGORITHM_SCRAM_SHA_1,
                         ScramDigestPassword.ALGORITHM_SCRAM_SHA_256
                 ))
@@ -469,38 +467,6 @@ class JdbcRealmDefinition extends SimpleResourceDefinition {
                 .setMinSize(1)
                 .setAllowDuplicates(true)
                 .build();
-    }
-
-    /**
-     * A simple {@link ModelTypeValidator} that requires that values are contained on a pre-defined list of string.
-     *
-     * //TODO: couldn't find a built-in validator for that. see if there is one or even if it can be moved to its own file.
-     */
-    static class StringValuesValidator extends ModelTypeValidator implements AllowedValuesValidator {
-
-        private List<ModelNode> allowedValues = new ArrayList<>();
-
-        StringValuesValidator(String... values) {
-            super(ModelType.STRING);
-            for (String value : values) {
-                allowedValues.add(new ModelNode().set(value));
-            }
-        }
-
-        @Override
-        public void validateParameter(String parameterName, ModelNode value) throws OperationFailedException {
-            super.validateParameter(parameterName, value);
-            if (value.isDefined()) {
-                if (!allowedValues.contains(value)) {
-                    throw new OperationFailedException(ControllerLogger.ROOT_LOGGER.invalidValue(value.asString(), parameterName, allowedValues));
-                }
-            }
-        }
-
-        @Override
-        public List<ModelNode> getAllowedValues() {
-            return this.allowedValues;
-        }
     }
 
     static final AttributeDefinition[] ATTRIBUTES = new AttributeDefinition[] {PrincipalQueryAttributes.PRINCIPAL_QUERIES, RealmDefinitions.CASE_SENSITIVE};
